feat(nvim): weekend maintenance

+++ private_dot_config/exact_nvim/exact_lua/exact_spec/exact_mini/readonly_mini-basics.lua
@@ -7,7 +7,7 @@ return {
     options = {
       basic = false,
       extra_ui = false,
-      win_borders = [[double]]
+      win_borders = [[single]]
     },
 
     mappings = {
