+++ private_dot_config/exact_nvim/exact_lua/exact_spec/exact_mini/readonly_mini-files.lua
@@ -24,7 +24,7 @@ return {
       pattern = [[MiniFilesWindowOpen]],
       callback = function(args)
         local win_id = args.data.win_id
-        vim.api.nvim_win_set_config(win_id, { border = [[double]] })
+        vim.api.nvim_win_set_config(win_id, { border = [[single]] })
       end
     })
